Move the Social settings endpoint to publicze
