+++ projects/plugins/social/src/class-rest-settings-controller.php
@@ -8,7 +8,6 @@
 
 namespace Automattic\Jetpack\Social;
 
-use Automattic\Jetpack\Modules;
 use Jetpack_Social;
 use WP_Error;
 use WP_REST_Controller;
@@ -26,24 +25,6 @@ class REST_Settings_Controller extends WP_REST_Controller {
 	 * @static
 	 */
 	public function register_rest_routes() {
-		register_rest_route(
-			'jetpack/v4',
-			'/social/settings',
-			array(
-				array(
-					'methods'             => WP_REST_Server::READABLE,
-					'callback'            => array( $this, 'get_item' ),
-					'permission_callback' => array( $this, 'require_admin_privilege_callback' ),
-					'args'                => $this->get_endpoint_args_for_item_schema(),
-				),
-				array(
-					'methods'             => WP_REST_Server::EDITABLE,
-					'callback'            => array( $this, 'update_item' ),
-					'permission_callback' => array( $this, 'require_admin_privilege_callback' ),
-					'args'                => $this->get_endpoint_args_for_item_schema( WP_REST_Server::EDITABLE ),
-				),
-			)
-		);
 		register_rest_route(
 			'jetpack/v4',
 			'/social/review-dismiss',
@@ -90,50 +71,6 @@ class REST_Settings_Controller extends WP_REST_Controller {
 			esc_html__( 'You are not allowed to perform this action.', 'jetpack-social' ),
 			array( 'status' => rest_authorization_required_code() )
 		);
-	}
-
-	/**
-	 * Updates the settings.
-	 *
-	 * @param WP_REST_Request $request Full details about the request.
-	 * @return array|WP_Error Array on success, or error object on failure.
-	 */
-	public function get_item( $request ) {
-		$fields = $this->get_fields_for_response( $request );
-		$data   = array();
-
-		if ( rest_is_field_included( 'publicize_active', $fields ) ) {
-			$data['publicize_active'] = Jetpack_Social::is_publicize_active();
-		}
-
-		return $this->prepare_item_for_response( $data, $request );
-	}
-
-	/**
-	 * POST `jetpack/v4/social/settings`
-	 *
-	 * @param WP_REST_Request $request - REST request.
-	 */
-	public function update_item( $request ) {
-		$params   = $request->get_params();
-		$settings = $this->get_endpoint_args_for_item_schema( $request->get_method() );
-
-		foreach ( array_keys( $settings ) as $name ) {
-			if ( ! array_key_exists( $name, $params ) ) {
-				continue;
-			}
-
-			switch ( $name ) {
-				case 'publicize_active':
-					$updated = ( new Modules() )->update_status( \Jetpack_Social::JETPACK_PUBLICIZE_MODULE_SLUG, (bool) $params[ $name ], false, false );
-					if ( is_wp_error( $updated ) ) {
-						return $updated;
-					}
-					break;
-			}
-		}
-
-		return $this->get_item( $request );
 	}
 
 	/**
@@ -184,26 +121,5 @@ class REST_Settings_Controller extends WP_REST_Controller {
 			$return[ $name ] = $sanitized;
 		}
 		return rest_ensure_response( $return );
-	}
-
-	/**
-	 * Get the settings schema, conforming to JSON Schema.
-	 *
-	 * @return array
-	 */
-	public function get_item_schema() {
-		$schema = array(
-			'$schema'    => 'http://json-schema.org/draft-04/schema#',
-			'title'      => 'system_status',
-			'type'       => 'object',
-			'properties' => array(
-				'publicize_active' => array(
-					'description' => __( 'Is the publicize module enabled?', 'jetpack-social' ),
-					'type'        => 'boolean',
-					'context'     => array( 'view', 'edit' ),
-				),
-			),
-		);
-		return $this->add_additional_fields_schema( $schema );
 	}
 }
